+++ Engine/VisualEffects/ParticleSystem/SphereEmiter.cpp
@@ -48,7 +48,7 @@ glm::vec3 SphereEmiter::GetParticlePosition ()
 {
 	glm::vec3 position;
 
-	float azimuthal = Random::Instance ().RangeF (-M_PI, M_PI);
+	float azimuthal = Random::Instance ().RangeF ((float) -M_PI, (float) M_PI);
 	float polar = acos (2 * Random::Instance ().RangeF (0, 1) - 1);
 	float radial = _radius * cbrt (Random::Instance ().RangeF (0, 1));
